Fixed lint issues Whitelisted folder containing sandbox testing files
