+++ lib/decorators.rb
@@ -1,6 +1,5 @@
 require_relative 'nameable'
 
-
 class Decorator < Nameable
   def initialize(nameable)
     super()
