+++ lib/rentals.rb
@@ -9,5 +9,4 @@ class Rental
     @book.rentals << self
     @person.rentals << self
   end
-
 end
